Inflation routine WIP.

+++ splash-dao-offchain/src/time.rs
@@ -4,10 +4,10 @@ use crate::constants::EPOCH_LEN;
 use crate::GenesisEpochStartTime;
 
 pub type NetworkTime = u64;
-pub type ProtocolEpoch = u64;
+pub type ProtocolEpoch = u32;
 
 pub fn epoch_start(gen_epoch_start: GenesisEpochStartTime, epoch: ProtocolEpoch) -> NetworkTime {
-    <u64>::from(gen_epoch_start) + epoch * EPOCH_LEN
+    <u64>::from(gen_epoch_start) + (epoch as u64) * EPOCH_LEN
 }
 
 pub fn epoch_end(gen_epoch_start: GenesisEpochStartTime, epoch: ProtocolEpoch) -> NetworkTime {
